Fixed continuous range calculations
It took some time for me to understand that the spaces between
successive colors are the buckets. So the original code had a off by
one error.

Also improved the battery color shading to my eye. Battery group will
stay green from roughly 60% to 100% and change to worrying shades
(first red, then when it is really important that you seek external
power purple) after that.